+++ src/js/main.js
@@ -1,4 +1,5 @@
 import { accordion } from './accordion';
+import { feedback } from './feedback';
 import { siteSearch } from './search';
 import { setup } from './setup';
 import { view } from './view';
@@ -10,20 +11,6 @@ window.addEventListener('load', () => {
     if (state !== 'grid') {
         accordion();
     }
+    feedback();
     setup();
-
-    const fbClose = document.querySelector('#feedback-close');
-    const fbOpen = document.querySelector('#feedback-open');
-    const fbOverlay = document.querySelector('#feedback-overlay');
-
-    fbClose.addEventListener('click', (e) => {
-        e.preventDefault();
-        fbOverlay.classList.toggle('feedback-overlay--active');
-    });
-    fbOpen.addEventListener('click', (e) => {
-        e.preventDefault();
-        fbOverlay.classList.toggle('feedback-overlay--active');
-        const fbInput = document.querySelector('#message');
-        fbInput.focus();
-    });
 });
